feat: remove flow and custom hook for PageWithStackHeader
