+++ cardstack/src/components/PageWithStackHeader/index.ts
@@ -1,2 +1,1 @@
 export { default as PageWithStackHeader } from './PageWithStackHeader';
-export { default as usePageWithStackHeader } from './usePageWithStackHeader';
